requirements.txt
mysql-connector-python==8.0.24
opencv-python==4.3.0.36
opencv-contrib-python==4.5.1.48
Pillow==8.0.1
torch==1.7.1